% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomous.h.R
\name{dichotomous}
\alias{dichotomous}
\title{Dichotomous Model}
\usage{
dichotomous(
  data,
  vars,
  itotal = TRUE,
  prop = TRUE,
  imeasure = FALSE,
  ise = FALSE,
  infit = FALSE,
  outfit = FALSE,
  reliability = TRUE,
  modelfit = FALSE,
  modelfitp = FALSE,
  wrightmap = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{itotal}{.}

\item{prop}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{reliability}{.}

\item{modelfit}{.}

\item{modelfitp}{.}

\item{wrightmap}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$scale} \tab \tab \tab \tab \tab a table \cr
  \code{results$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$scale$asDF}

\code{as.data.frame(results$scale)}
}
\description{
Dichotomous Model
}
