% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.h.R
\name{facet}
\alias{facet}
\title{Many Facet Rasch Model}
\usage{
facet(
  data,
  dep,
  id,
  facet,
  time,
  tes = FALSE,
  plot9 = FALSE,
  driftm = FALSE,
  driftfit = FALSE,
  rm = FALSE,
  im = FALSE,
  inter = FALSE,
  raw = FALSE,
  sm = FALSE,
  ifit = FALSE,
  pm = FALSE,
  pfit = FALSE,
  resid = FALSE,
  pca = FALSE,
  local = FALSE,
  plot1 = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  angle = 0,
  plot4 = FALSE,
  plot5 = FALSE,
  num = 1,
  plot6 = FALSE,
  num1 = 1,
  plot7 = FALSE,
  angle1 = 0,
  plot8 = FALSE
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{id}{.}

\item{facet}{.}

\item{time}{.}

\item{tes}{.}

\item{plot9}{.}

\item{driftm}{.}

\item{driftfit}{.}

\item{rm}{.}

\item{im}{.}

\item{inter}{.}

\item{raw}{.}

\item{sm}{.}

\item{ifit}{.}

\item{pm}{.}

\item{pfit}{.}

\item{resid}{.}

\item{pca}{.}

\item{local}{.}

\item{plot1}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{angle}{.}

\item{plot4}{.}

\item{plot5}{.}

\item{num}{.}

\item{plot6}{.}

\item{num1}{.}

\item{plot7}{.}

\item{angle1}{.}

\item{plot8}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$rm} \tab \tab \tab \tab \tab a table \cr
  \code{results$im} \tab \tab \tab \tab \tab a table \cr
  \code{results$sm} \tab \tab \tab \tab \tab a table \cr
  \code{results$raw} \tab \tab \tab \tab \tab a table \cr
  \code{results$inter} \tab \tab \tab \tab \tab a table \cr
  \code{results$ifit} \tab \tab \tab \tab \tab a table \cr
  \code{results$driftm} \tab \tab \tab \tab \tab a table \cr
  \code{results$driftfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$driftsum} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot9} \tab \tab \tab \tab \tab an image \cr
  \code{results$pm} \tab \tab \tab \tab \tab a table \cr
  \code{results$pfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$resid} \tab \tab \tab \tab \tab a table \cr
  \code{results$pca} \tab \tab \tab \tab \tab a table \cr
  \code{results$local} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot7} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot8} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$rm$asDF}

\code{as.data.frame(results$rm)}
}
\description{
Many Facet Rasch Model
}
