% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.h.R
\name{item}
\alias{item}
\title{Distractor analysis}
\usage{
item(
  data,
  vars,
  key = "D,C,A,D,D,A,D,B,D,A,A,D,C,C,B,C,D,A,A,B",
  num = 1,
  count = TRUE,
  prop = FALSE,
  plot = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{key}{.}

\item{num}{.}

\item{count}{.}

\item{prop}{.}

\item{plot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$count} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$prop} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Distractor analysis
}
