% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytomous.h.R
\name{polytomous}
\alias{polytomous}
\title{Polytomous Model}
\usage{
polytomous(
  data,
  vars,
  imeasure = TRUE,
  ise = TRUE,
  infit = FALSE,
  outfit = FALSE,
  reliability = TRUE,
  modelfit = FALSE,
  modelfitp = FALSE,
  mat = FALSE,
  thresh = TRUE,
  pmeasure = FALSE,
  thurs = FALSE,
  wplot = TRUE,
  inplot = FALSE,
  outplot = FALSE,
  angle = 0,
  tau = FALSE,
  model = FALSE,
  lr = FALSE,
  piplot = TRUE,
  to = FALSE,
  st = FALSE,
  plot2 = FALSE,
  plot3 = FALSE,
  width = 500,
  height = 500,
  width5 = 500,
  height5 = 500,
  plot4 = FALSE,
  plot6 = FALSE,
  obs = "TRUE",
  width7 = 500,
  height7 = 500,
  width3 = 500,
  height3 = 500,
  width2 = 500,
  height2 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{reliability}{.}

\item{modelfit}{.}

\item{modelfitp}{.}

\item{mat}{.}

\item{thresh}{.}

\item{pmeasure}{.}

\item{thurs}{.}

\item{wplot}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{tau}{.}

\item{model}{.}

\item{lr}{.}

\item{piplot}{.}

\item{to}{.}

\item{st}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{width}{.}

\item{height}{.}

\item{width5}{.}

\item{height5}{.}

\item{plot4}{.}

\item{plot6}{.}

\item{obs}{.}

\item{width7}{.}

\item{height7}{.}

\item{width3}{.}

\item{height3}{.}

\item{width2}{.}

\item{height2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$ia$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$ia$thresh} \tab \tab \tab \tab \tab a table \cr
  \code{results$ia$thurs} \tab \tab \tab \tab \tab a table \cr
  \code{results$mf$scale} \tab \tab \tab \tab \tab a table \cr
  \code{results$mf$mat} \tab \tab \tab \tab \tab a table \cr
  \code{results$mcc$model} \tab \tab \tab \tab \tab a table \cr
  \code{results$mcc$lr} \tab \tab \tab \tab \tab a table \cr
  \code{results$ss$st} \tab \tab \tab \tab \tab a table \cr
  \code{results$ss$to} \tab \tab \tab \tab \tab a table \cr
  \code{results$wplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$piplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an array of images \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an array of images \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$total} \tab \tab \tab \tab \tab an output \cr
  \code{results$personmeasure} \tab \tab \tab \tab \tab an output \cr
  \code{results$pse} \tab \tab \tab \tab \tab an output \cr
  \code{results$pinfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$poutfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$resid} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}
}
\description{
Polytomous Model
}
