% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.h.R
\name{clr}
\alias{clr}
\title{Conditional Likelihood Ratio Test}
\usage{
clr(data, vars, group, model, clr = TRUE, resi = FALSE, score = "low")
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{A string naming the grouping variable from \code{data}}

\item{model}{.}

\item{clr}{.}

\item{resi}{.}

\item{score}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$clr} \tab \tab \tab \tab \tab a table \cr
  \code{results$resi} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$clr$asDF}

\code{as.data.frame(results$clr)}
}
\description{
Conditional Likelihood Ratio Test
}
