% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.h.R
\name{clr}
\alias{clr}
\title{Conditional Likelihood Ratio Test}
\usage{
clr(
  data,
  vars,
  group,
  model,
  num = 1,
  ci = 3,
  clr = TRUE,
  resi = FALSE,
  score,
  dif = FALSE,
  plot = FALSE,
  plot1 = FALSE,
  plot2 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{A string naming the grouping variable from \code{data}}

\item{model}{.}

\item{num}{.}

\item{ci}{.}

\item{clr}{.}

\item{resi}{.}

\item{score}{.}

\item{dif}{.}

\item{plot}{.}

\item{plot1}{.}

\item{plot2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$clr} \tab \tab \tab \tab \tab a table \cr
  \code{results$resi} \tab \tab \tab \tab \tab a table \cr
  \code{results$dif} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$clr$asDF}

\code{as.data.frame(results$clr)}
}
\description{
Conditional Likelihood Ratio Test
}
