% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltam.h.R
\name{deltam}
\alias{deltam}
\title{Delta method}
\usage{
deltam(
  data,
  vars,
  group,
  fixed = TRUE,
  plot = FALSE,
  normal = FALSE,
  puri,
  plot1 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{A string naming the grouping variable from \code{data}}

\item{fixed}{.}

\item{plot}{.}

\item{normal}{.}

\item{puri}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$fixed} \tab \tab \tab \tab \tab a table \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$normal} \tab \tab \tab \tab \tab a table \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$fixed$asDF}

\code{as.data.frame(results$fixed)}
}
\description{
Delta method
}
