% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dichotomous.h.R
\name{dichotomous}
\alias{dichotomous}
\title{Dichotomous Rasch Model}
\usage{
dichotomous(
  data,
  vars,
  prop = TRUE,
  imeasure = FALSE,
  ise = FALSE,
  infit = FALSE,
  outfit = FALSE,
  reliability = TRUE,
  modelfit = FALSE,
  modelfitp = FALSE,
  mat = FALSE,
  wrightmap = TRUE,
  esc = FALSE,
  inplot = FALSE,
  outplot = FALSE,
  angle = 0,
  to = FALSE,
  plot2 = FALSE,
  st = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{prop}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{reliability}{.}

\item{modelfit}{.}

\item{modelfitp}{.}

\item{mat}{.}

\item{wrightmap}{.}

\item{esc}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{to}{.}

\item{plot2}{.}

\item{st}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$st} \tab \tab \tab \tab \tab a table \cr
  \code{results$to} \tab \tab \tab \tab \tab a table \cr
  \code{results$scale} \tab \tab \tab \tab \tab a table \cr
  \code{results$mat} \tab \tab \tab \tab \tab a table \cr
  \code{results$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$esc} \tab \tab \tab \tab \tab an array of plots \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$total} \tab \tab \tab \tab \tab an output \cr
  \code{results$pmeasure} \tab \tab \tab \tab \tab an output \cr
  \code{results$pse} \tab \tab \tab \tab \tab an output \cr
  \code{results$pinfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$poutfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$resid} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$st$asDF}

\code{as.data.frame(results$st)}
}
\description{
Dichotomous Rasch Model
}
