% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.h.R
\name{dif}
\alias{dif}
\title{Raju & Mantel-Haenszel method}
\usage{
dif(
  data,
  vars,
  group,
  raju = FALSE,
  mh = FALSE,
  padjust = "BH",
  padjust1 = "BH",
  zplot = FALSE,
  iccplot = FALSE,
  plot1 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{A string naming the grouping variable from \code{data}}

\item{raju}{.}

\item{mh}{.}

\item{padjust}{.}

\item{padjust1}{.}

\item{zplot}{.}

\item{iccplot}{.}

\item{plot1}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$raju} \tab \tab \tab \tab \tab a table \cr
  \code{results$mh} \tab \tab \tab \tab \tab a table \cr
  \code{results$zplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$iccplot} \tab \tab \tab \tab \tab an array of plots \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$raju$asDF}

\code{as.data.frame(results$raju)}
}
\description{
Raju & Mantel-Haenszel method
}
