% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dif.h.R
\name{dif}
\alias{dif}
\title{Differential Item Functioning}
\usage{
dif(data, vars, group, raju = TRUE, difplot = FALSE)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{A string naming the grouping variable from \code{data}}

\item{raju}{.}

\item{difplot}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$raju} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$raju$asDF}

\code{as.data.frame(results$raju)}
}
\description{
Differential Item Functioning
}
