% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.h.R
\name{facet}
\alias{facet}
\title{Many Facet Rasch Model}
\usage{
facet(
  data,
  dep,
  id,
  facet,
  rm = FALSE,
  im = FALSE,
  inter = FALSE,
  rs = FALSE,
  sm = FALSE,
  ifit = FALSE,
  sifit = FALSE,
  pm = FALSE,
  pfit = FALSE,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500,
  plot2 = FALSE,
  width2 = 500,
  height2 = 500,
  plot3 = FALSE,
  width3 = 500,
  height3 = 500,
  angle = 0,
  plot5 = FALSE,
  width5 = 500,
  height5 = 500,
  angle1 = 0,
  plot4 = FALSE,
  width4 = 500,
  height4 = 500
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{id}{.}

\item{facet}{.}

\item{rm}{.}

\item{im}{.}

\item{inter}{.}

\item{rs}{.}

\item{sm}{.}

\item{ifit}{.}

\item{sifit}{.}

\item{pm}{.}

\item{pfit}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}

\item{plot2}{.}

\item{width2}{.}

\item{height2}{.}

\item{plot3}{.}

\item{width3}{.}

\item{height3}{.}

\item{angle}{.}

\item{plot5}{.}

\item{width5}{.}

\item{height5}{.}

\item{angle1}{.}

\item{plot4}{.}

\item{width4}{.}

\item{height4}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$rm} \tab \tab \tab \tab \tab a table \cr
  \code{results$im} \tab \tab \tab \tab \tab a table \cr
  \code{results$inter} \tab \tab \tab \tab \tab a table \cr
  \code{results$rs} \tab \tab \tab \tab \tab a table \cr
  \code{results$sm} \tab \tab \tab \tab \tab a table \cr
  \code{results$ifit} \tab \tab \tab \tab \tab a table \cr
  \code{results$sifit} \tab \tab \tab \tab \tab a table \cr
  \code{results$pm} \tab \tab \tab \tab \tab a table \cr
  \code{results$pfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$rm$asDF}

\code{as.data.frame(results$rm)}
}
\description{
Many Facet Rasch Model
}
