% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.h.R
\name{facet}
\alias{facet}
\title{Facet Model}
\usage{
facet(
  data,
  dep,
  id,
  facet,
  rm = FALSE,
  im = FALSE,
  inter = FALSE,
  sm = FALSE,
  ifit = FALSE
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{id}{.}

\item{facet}{.}

\item{rm}{.}

\item{im}{.}

\item{inter}{.}

\item{sm}{.}

\item{ifit}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$im} \tab \tab \tab \tab \tab a table \cr
  \code{results$rm} \tab \tab \tab \tab \tab a table \cr
  \code{results$inter} \tab \tab \tab \tab \tab a table \cr
  \code{results$sm} \tab \tab \tab \tab \tab a table \cr
  \code{results$ifit} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$im$asDF}

\code{as.data.frame(results$im)}
}
\description{
Facet Model
}
