% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.h.R
\name{facet}
\alias{facet}
\title{Many Facet Rasch Model}
\usage{
facet(
  data,
  dep,
  id,
  facet,
  rm = FALSE,
  im = FALSE,
  inter = FALSE,
  raw = FALSE,
  sm = FALSE,
  ifit = FALSE,
  pm = FALSE,
  pfit = FALSE,
  plot1 = FALSE,
  width1 = 500,
  height1 = 500,
  plot2 = FALSE,
  width2 = 500,
  height2 = 500,
  plot3 = FALSE,
  width3 = 500,
  height3 = 500,
  angle = 0,
  plot4 = FALSE,
  width4 = 500,
  height4 = 500,
  plot5 = FALSE,
  num = 1,
  width5 = 500,
  height5 = 500,
  plot6 = FALSE,
  num1 = 1,
  width6 = 500,
  height6 = 500,
  plot7 = FALSE,
  width7 = 500,
  height7 = 500,
  angle1 = 0,
  plot8 = FALSE,
  width8 = 500,
  height8 = 500,
  g = FALSE,
  d = FALSE,
  formula = "value ~ (1 | subject) + (1 | task) + (1 | rater:task) + (1 | subject:task)"
)
}
\arguments{
\item{data}{.}

\item{dep}{.}

\item{id}{.}

\item{facet}{.}

\item{rm}{.}

\item{im}{.}

\item{inter}{.}

\item{raw}{.}

\item{sm}{.}

\item{ifit}{.}

\item{pm}{.}

\item{pfit}{.}

\item{plot1}{.}

\item{width1}{.}

\item{height1}{.}

\item{plot2}{.}

\item{width2}{.}

\item{height2}{.}

\item{plot3}{.}

\item{width3}{.}

\item{height3}{.}

\item{angle}{.}

\item{plot4}{.}

\item{width4}{.}

\item{height4}{.}

\item{plot5}{.}

\item{num}{.}

\item{width5}{.}

\item{height5}{.}

\item{plot6}{.}

\item{num1}{.}

\item{width6}{.}

\item{height6}{.}

\item{plot7}{.}

\item{width7}{.}

\item{height7}{.}

\item{angle1}{.}

\item{plot8}{.}

\item{width8}{.}

\item{height8}{.}

\item{g}{.}

\item{d}{.}

\item{formula}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text1} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text2} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$rm} \tab \tab \tab \tab \tab a table \cr
  \code{results$im} \tab \tab \tab \tab \tab a table \cr
  \code{results$sm} \tab \tab \tab \tab \tab a table \cr
  \code{results$raw} \tab \tab \tab \tab \tab a table \cr
  \code{results$inter} \tab \tab \tab \tab \tab a table \cr
  \code{results$ifit} \tab \tab \tab \tab \tab a table \cr
  \code{results$pm} \tab \tab \tab \tab \tab a table \cr
  \code{results$pfit} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot7} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot8} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot4} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot5} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot6} \tab \tab \tab \tab \tab an image \cr
  \code{results$g} \tab \tab \tab \tab \tab a table \cr
  \code{results$d} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$rm$asDF}

\code{as.data.frame(results$rm)}
}
\description{
Many Facet Rasch Model
}
