% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item.h.R
\name{item}
\alias{item}
\title{Distractor analysis}
\usage{
item(
  data,
  vars,
  key = "D,C,A,D,D,A,D,B,D,A,A,D,C,C,B,C,D,A,A,B",
  num = 1,
  num1 = 1,
  group = 3,
  group1 = 3,
  count = TRUE,
  prop = FALSE,
  sum = FALSE,
  disc = FALSE,
  sum1 = FALSE,
  plot = FALSE,
  angle = 0,
  plot1 = FALSE,
  disi,
  plot2 = FALSE,
  plot3 = FALSE,
  width2 = 500,
  height2 = 500,
  width = 500,
  height = 500,
  width1 = 500,
  height1 = 500,
  width3 = 500,
  height3 = 500
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{key}{.}

\item{num}{.}

\item{num1}{.}

\item{group}{.}

\item{group1}{.}

\item{count}{.}

\item{prop}{.}

\item{sum}{.}

\item{disc}{.}

\item{sum1}{.}

\item{plot}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{plot1}{.}

\item{disi}{.}

\item{plot2}{.}

\item{plot3}{.}

\item{width2}{.}

\item{height2}{.}

\item{width}{.}

\item{height}{.}

\item{width1}{.}

\item{height1}{.}

\item{width3}{.}

\item{height3}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$count} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$prop} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$sum} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$disc} \tab \tab \tab \tab \tab a table \cr
  \code{results$sum1} \tab \tab \tab \tab \tab an array of tables \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
  \code{results$plot1} \tab \tab \tab \tab \tab an image \cr
  \code{results$total} \tab \tab \tab \tab \tab an output \cr
  \code{results$scoring} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot3} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$disc$asDF}

\code{as.data.frame(results$disc)}
}
\description{
Distractor analysis
}
