% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.h.R
\name{logit}
\alias{logit}
\title{DIF for ordinal data}
\usage{
logit(
  data,
  vars,
  anchor,
  group,
  model = "adjacent",
  type = "both",
  match = "zscore",
  padjust = "BH",
  method = FALSE,
  plot = FALSE,
  plotItem = 1,
  anchorMethod = "all",
  plotLayout = "faceted"
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{anchor}{.}

\item{group}{A string naming the grouping variable from \code{data}}

\item{model}{.}

\item{type}{.}

\item{match}{.}

\item{padjust}{.}

\item{method}{.}

\item{plot}{.}

\item{plotItem}{.}

\item{anchorMethod}{.}

\item{plotLayout}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$method} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$method$asDF}

\code{as.data.frame(results$method)}
}
\description{
DIF for ordinal data
}
