% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.h.R
\name{logit}
\alias{logit}
\title{DIF with cumulative logit}
\usage{
logit(
  data,
  vars,
  group,
  model = "adjacent",
  type = "both",
  match = "zscore",
  padjust = "none",
  dif = TRUE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{group}{A string naming the grouping variable from \code{data}}

\item{model}{.}

\item{type}{.}

\item{match}{.}

\item{padjust}{.}

\item{dif}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$dif} \tab \tab \tab \tab \tab a table \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$dif$asDF}

\code{as.data.frame(results$dif)}
}
\description{
DIF with cumulative logit
}
