% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytomous.h.R
\name{polytomous}
\alias{polytomous}
\title{Polytomous Rasch Model}
\usage{
polytomous(
  data,
  vars,
  imeasure = TRUE,
  ise = TRUE,
  infit = FALSE,
  outfit = FALSE,
  reliability = TRUE,
  modelfit = FALSE,
  modelfitp = FALSE,
  mat = FALSE,
  thresh = FALSE,
  pmeasure = FALSE,
  icc = FALSE,
  wrightmap = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{reliability}{.}

\item{modelfit}{.}

\item{modelfitp}{.}

\item{mat}{.}

\item{thresh}{.}

\item{pmeasure}{.}

\item{icc}{.}

\item{wrightmap}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$scale} \tab \tab \tab \tab \tab a table \cr
  \code{results$mat} \tab \tab \tab \tab \tab a table \cr
  \code{results$thresh} \tab \tab \tab \tab \tab a table \cr
  \code{results$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$plot} \tab \tab \tab \tab \tab an array of plots \cr
  \code{results$wrightmap} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$scale$asDF}

\code{as.data.frame(results$scale)}
}
\description{
Polytomous Rasch Model
}
