% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytomous.h.R
\name{polytomous}
\alias{polytomous}
\title{Polytomous Rasch Model}
\usage{
polytomous(
  data,
  vars,
  imeasure = TRUE,
  ise = TRUE,
  infit = FALSE,
  outfit = FALSE,
  reliability = TRUE,
  modelfit = FALSE,
  modelfitp = FALSE,
  mat = FALSE,
  thresh = TRUE,
  pmeasure = FALSE,
  thurs = FALSE,
  icc = FALSE,
  wrightmap = TRUE,
  esc = FALSE,
  inplot = FALSE,
  outplot = FALSE,
  angle = 0,
  tau = FALSE,
  model = FALSE,
  lr = FALSE,
  piplot = TRUE,
  to = FALSE,
  st = FALSE,
  plot2 = FALSE
)
}
\arguments{
\item{data}{The data as a data frame.}

\item{vars}{.}

\item{imeasure}{.}

\item{ise}{.}

\item{infit}{.}

\item{outfit}{.}

\item{reliability}{.}

\item{modelfit}{.}

\item{modelfitp}{.}

\item{mat}{.}

\item{thresh}{.}

\item{pmeasure}{.}

\item{thurs}{.}

\item{icc}{.}

\item{wrightmap}{.}

\item{esc}{.}

\item{inplot}{.}

\item{outplot}{.}

\item{angle}{a number from 0 to 90 defining the angle of the x-axis labels,
where 0 degrees represents completely horizontal labels.}

\item{tau}{.}

\item{model}{.}

\item{lr}{.}

\item{piplot}{.}

\item{to}{.}

\item{st}{.}

\item{plot2}{.}
}
\value{
A results object containing:
\tabular{llllll}{
  \code{results$instructions} \tab \tab \tab \tab \tab a html \cr
  \code{results$st} \tab \tab \tab \tab \tab a table \cr
  \code{results$to} \tab \tab \tab \tab \tab a table \cr
  \code{results$scale} \tab \tab \tab \tab \tab a table \cr
  \code{results$model} \tab \tab \tab \tab \tab a table \cr
  \code{results$lr} \tab \tab \tab \tab \tab a table \cr
  \code{results$mat} \tab \tab \tab \tab \tab a table \cr
  \code{results$thresh} \tab \tab \tab \tab \tab a table \cr
  \code{results$thurs} \tab \tab \tab \tab \tab a table \cr
  \code{results$items} \tab \tab \tab \tab \tab a table \cr
  \code{results$wrightmap} \tab \tab \tab \tab \tab an image \cr
  \code{results$piplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$esc} \tab \tab \tab \tab \tab an array of plots \cr
  \code{results$plot} \tab \tab \tab \tab \tab an array of plots \cr
  \code{results$inplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$outplot} \tab \tab \tab \tab \tab an image \cr
  \code{results$total} \tab \tab \tab \tab \tab an output \cr
  \code{results$personmeasure} \tab \tab \tab \tab \tab an output \cr
  \code{results$pse} \tab \tab \tab \tab \tab an output \cr
  \code{results$pinfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$poutfit} \tab \tab \tab \tab \tab an output \cr
  \code{results$text} \tab \tab \tab \tab \tab a preformatted \cr
  \code{results$resid} \tab \tab \tab \tab \tab an output \cr
  \code{results$plot2} \tab \tab \tab \tab \tab an image \cr
}

Tables can be converted to data frames with \code{asDF} or \code{\link{as.data.frame}}. For example:

\code{results$st$asDF}

\code{as.data.frame(results$st)}
}
\description{
Polytomous Rasch Model
}
